Implemented MULT and DIV SINASM instructions, sign checking, bug fixes
Implemented mult/multu and div/divu instructions so we don't need builtins functions for them. Further, implemented a function in the compiler to determine whether data is signed, and whether we should use the signed or unsigend versions of instructions. There were also bug fixes in the parse and compiler relating to empty function definitions. The builtins library was updated to reflect all of these changes.
Further, compiler symbols now use a vector of qualities so you aren't limited to just one. Note some qualities will conflict, but these conflicts have not yet been addressed, nor have the qualities relating to static/dynamic memory.

+++ SINVM.h
@@ -70,7 +70,7 @@ class SINVM
 	std::vector<uint8_t> get_properly_ordered_bytes(int value);
 
 	// execute a single instruction
-	void execute_instruction(int opcode);
+	void execute_instruction(uint16_t opcode);
 
 	// instruction-specific load/store functions
 	uint16_t execute_load();
